split up tokenization related
